pyproject.toml
[tool.poetry]
name = "topocluster"
authors = ["Myles Bartlett"]
description = "Fine-tuning SSL Representations with Algebraic Topology"
version = "0.0.1"

[tool.poetry.dependencies]
python = "3.8"
wandb = "^0.10.8"
scipy = "^1.5.3"
lapjv = "^1.3.1"
pykeops = "^1.4.1"
gudhi = "^3.3.0"
eagerpy = "^0.29.0"
POT = "^0.7.0"
pytorch-lightning = "^1.4.1"
hydra-core = {git = "https://github.com/facebookresearch/hydra.git", branch= "main"}
matplotlib = "^3.3.2"
neoconfigen = {git = "https://github.com/predictive-analytics-lab/neoconfigen.git", branch="main"}
pandas = "^1.2.3"
black = "^20.8b1"
torch = "^1.8.0"
faiss-cpu = "^1.7.0"
torchvision = "^0.10.1"
numba = "^0.53.0"
scikit-learn = "^0.24.1"
conduit = {git = "https://github.com/predictive-analytics-lab/conduit.git", branch = "main"}
umap-learn = "^0.5.1"
pretrainedmodels = "^0.7.4"
fast-soft-sort = {git = "https://github.com/google-research/fast-soft-sort.git"}
palkit = {git = "https://github.com/predictive-analytics-lab/palkit.git", branch = "main" }
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"
mypy = "^0.790"
flake8 = "^3.8.4"

[tool.black]
line-length = 100
target-version = ['py38']
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
include_trailing_comma = "True"
known_third_party = [
  "black",
  "click",
  "click_option_group",
  "matplotlib",
  "mypy",
  "numpy",
  "pandas",
  "pylint",
  "pytest",
  "scipy",
  "setuptools",
  "sklearn",
  "torch",
  "tqdm",
]
line_length = 100
multi_line_output = 3


[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=1.0.10"]
